move a bunch of stray zarr-related items into their own files
